+++ covalve/runtime/nodes/__init__.py
@@ -1,0 +1,3 @@
+from covalve.runtime.nodes.registry import NodesRegistry
+
+node = NodesRegistry()